Updated types for addon actions
